last xpdf version from ctan.org

+++ goo/FixedPoint.cc
@@ -16,7 +16,6 @@
 #pragma implementation
 #endif
 
-#include "gmempp.h"
 #include "FixedPoint.h"
 
 #define ln2 ((FixedPoint)0.69314718)
@@ -114,7 +113,7 @@ GBool FixedPoint::divCheck(FixedPoint x, FixedPoint y, FixedPoint *result) {
       z >= ((FixPtInt64)1 << 31) || z < -((FixPtInt64)1 << 31)) {
     return gFalse;
   }
-  result->val = (int)z;
+  result->val = z;
   return gTrue;
 }
